Move part of the API code to the VirtusizeCore module
