+++ Virtusize/Sources/UI/VirtusizeWebViewController.swift
@@ -24,7 +24,6 @@
 
 import UIKit
 import WebKit
-import VirtusizeAuth
 
 /// The methods of this protocol notify you with Virtusize specific messages such as errors as
 /// `VirtusizeError` and events as `VirtusizeEvent`
