+++ Virtusize/Sources/Virtusize.h
@@ -24,6 +24,8 @@
 
 
 @import UIKit;
+@import VirtusizeCore;
+@import VirtusizeAuth;
 
 FOUNDATION_EXPORT double VirtusizeVersionNumber;
 FOUNDATION_EXPORT const unsigned char VirtusizeVersionString[];
